requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
Faker==18.7.0
flake8==6.0.0
flake8-builtins==2.1.0
flake8-django==1.2
flake8-import-order==0.18.2
flake8-print==5.0.0
idna==3.4
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
